+++ src/schemathesis/cli/__main__.py
@@ -1,0 +1,4 @@
+if __name__ == "__main__":
+    import schemathesis.cli
+
+    schemathesis.cli.schemathesis()